+++ src/main/java/com/ndb/auction/resolver/WalletResolver.java
@@ -1,11 +1,5 @@
 package com.ndb.auction.resolver;
 
-import java.util.List;
-
-import com.ndb.auction.models.InternalBalance;
-import com.ndb.auction.service.UserDetailsImpl;
-
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 
 import graphql.kickstart.tools.GraphQLMutationResolver;
@@ -38,7 +32,7 @@ public class WalletResolver extends BaseResolver implements GraphQLQueryResolver
         return ndbWalletService.getWalletBalance(token, network, address).toString();
     }
 
-    public Boolean transferFunds(String token, String network, String address, int amount) {
+    public Boolean transferFunds(String token, String network, String address, long amount) {
         return ndbWalletService.transferFunds(token, network, address, amount);
     }
